Create a separate inference deployment per detector  (#38)
* tests SDK tests from a different branch

* initial commit

* create detector deployment script

* [wip] got detector deployment working

* wip

* [wip] testing how long it takes to k apply a deployment

* [wip]

* working kubernetes client

* Automatically reformatting code with black and isort

* clean up

* more clean up

* better exception handling

* mount inference deployment template as a configmap

* Automatically reformatting code with black and isort

* code refactoring

* better comments

* miscellaneous improvements on edge inference

* better comments

* tests

* fix tests

* tests

* remove comments

* fixing tests

* Automatically reformatting code with black and isort

* address pr feedback

* addressing pr feedback

* tests

* Running inference with a CPU model on GHA's and various fixes  (#43)



* improvements

* Automatically reformatting code with black and isort

* implementation fix

* Automatically reformatting code with black and isort

* actual fixes


* Automatically reformatting code with black and isort

---------


* nvidia-runtime fix on g4  (#44)

* various fixes

* gha

* update image

* addressing pr feedback

* more fixes

* update image

* Automatically reformatting code with black and isort

* change log level to info

* fix tritonclient version

* fixing path issue

---------

+++ deploy/bin/install-k3s-g4.sh
@@ -16,9 +16,12 @@ SCRIPT_DIR=$(dirname "$0")
 $SCRIPT_DIR/install-k3s.sh
 
 # Configure k3s to use nvidia-container-runtime
-# See guide here: https://k3d.io/v5.3.0/usage/advanced/cuda/#configure-containerd
+# See guide here: https://k3d.io/v5.3.0/usage/advanced/cuda/#configure-containerd 
 echo "Configuring k3s to use nvidia-container-runtime..."
 for i in {1..10}; do
+  if [[ -f "/var/lib/rancher/k3s/agent/etc/containerd/config.toml.tmpl" ]]; then
+    break
+  fi
   # Sometimes the following wget fails initially but works after a few seconds
   sleep 2
   sudo wget https://k3d.io/v5.3.0/usage/advanced/cuda/config.toml.tmpl -O /var/lib/rancher/k3s/agent/etc/containerd/config.toml.tmpl && break
